requirements.txt
alembic==0.6.2
bleach==1.4
blinker==1.3
click==6.6
colorama==0.2.7
coverage==3.7.1
dominate==2.3.1
Flask==0.10.1
Flask-Admin==1.4.2
Flask-Bootstrap==3.0.3.1
Flask-HTTPAuth==2.7.0
Flask-Login==0.3.1
Flask-Mail==0.9.0
Flask-Migrate==1.1.0
Flask-Moment==0.2.1
Flask-PageDown==0.1.4
Flask-Script==0.6.6
Flask-SQLAlchemy==1.0
Flask-WTF==0.9.4
ForgeryPy==0.1
gunicorn==19.6.0
html5lib==1.0b3
httpie==0.7.2
image==1.5.5
itsdangerous==0.23
Jinja2==2.7.1
Mako==0.9.1
Markdown==2.3.1
MarkupSafe==0.18
numpy==1.11.3
Pillow==3.4.2
psycopg2==2.6.2
Pygments==1.6
python-editor==1.0.3
requests==2.1.0
selenium==2.45.0
six==1.4.1
SQLAlchemy==0.9.9
visitor==0.1.3
Werkzeug==0.10.4
WTForms==1.0.5
